backend/requirements.txt
flask
flask-cors
openai>=1.0.0
python-dotenv
azure-storage-blob